requirements.txt
pyspark
apache-airflow-providers-apache-spark
apache-airflow-providers-slack
apache-airflow-providers-postgres
great-expectations==1.3.11
boto3
pandas
apache-airflow-providers-amazon




